requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
anthropic>=0.7.0
httpx>=0.25.0
qrcode[pil]>=7.4.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
paho-mqtt>=2.0.0
prometheus-client>=0.19.0
locust>=2.17.0
redis>=5.0.0
